pyproject.toml
[project]
name = "nba26"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.2",
    "nba-api>=1.11.3",
    "pandas>=2.3.3",
    "snowflake-connector-python>=4.0.0",
    "sqlmesh>=0.227.1",
]
